+++ packages/panzer/adapters-stk/test/stk_interface_test/tSquareQuadMeshFactory.cpp
@@ -122,7 +122,7 @@ void edge_block_test_helper(Teuchos::FancyOStream &out,
    if (expected_edge_block_count == 1) {
       std::vector<stk::mesh::Entity> edges;
       mesh->getMyEdges(edges);
-      TEST_ASSERT(all_edge_blocks[0]->entity_count() == edges.size());
+      TEST_ASSERT(all_edge_blocks[0]->entity_count() == (int64_t)edges.size());
    }
    }
 }
